meson: Fix rst2html5 detection
The first part of the version string contains the name that the
rst2html5 command was invoked as, which can differ based on the
operating system: on FreeBSD, for example, it's rst2html5.py
instead of just rst2html5.

Fix our detection logic so that it works regardless of the
specific name used for the docutils-provided rst2html5 command.

Fixes: cf0c9e186565e886a0016b2b269088b3eed3d26d
Signed-off-by: Andrea Bolognani <abologna@redhat.com>